accel mouse has prefs, and uses util for server
I added the ability for users to change sensativity and threshold in
preferences, as well as it will save any changes made while using accel
activity. I also integrated the server code that was used in the Util
shortening the mouse mode. As far as initial testing as shown, no errors.